+++ compiler/plc_diagnostics/src/lib.rs
@@ -1,9 +1,6 @@
 //! This crates handles diagnostics, from creation to reporting.
 
+pub mod diagnostician;
 pub mod diagnostics;
-pub mod reporter {
-    pub mod clang;
-    pub mod codespan;
-    pub mod null;
-}
 pub mod errno; // TODO: Make this crate-private?
+pub mod reporter;
