+++ compiler/plc_driver/src/pipelines.rs
@@ -21,10 +21,7 @@ use plc::{
     validation::Validator,
     ConfigFormat, Target,
 };
-use plc_diagnostics::{
-    diagnostics::{Diagnostic, Diagnostician},
-    errno::ErrNo,
-};
+use plc_diagnostics::{diagnostician::Diagnostician, diagnostics::Diagnostic, errno::ErrNo};
 use project::{
     object::Object,
     project::{LibraryInformation, Project},
