+++ compiler/plc_driver/src/tests.rs
@@ -2,7 +2,7 @@ use std::{fmt::Debug, path::PathBuf};
 
 use ast::provider::IdProvider;
 use plc::DebugLevel;
-use plc_diagnostics::diagnostics::{Diagnostic, Diagnostician};
+use plc_diagnostics::{diagnostician::Diagnostician, diagnostics::Diagnostic};
 use project::project::Project;
 use source_code::SourceContainer;
